requirements.txt
# Core Flask
Flask>=2.3.3
Flask-SQLAlchemy>=3.0.5
Flask-Login>=0.6.3
Flask-WTF>=1.1.1
Werkzeug>=2.3.7

# Authentication
authlib>=1.2.1
requests-oauthlib>=1.3.1
bcrypt>=4.0.1

# APIs
requests>=2.31.0
groq>=0.3.0

# Utilities - Let pip choose compatible versions
Pillow>=10.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
WTForms>=3.0.0
email-validator>=2.0.0
psycopg2-binary>=2.9.7

# Production
gunicorn>=21.2.0